[Refact] refactoring된 useFetch 적용
상세 페이지에 리팩토링된 useFetch 적용.
