+++ web/src/containers/PostDetailPage/PostSideBox/reducer.js
@@ -1,6 +1,7 @@
 import produce from 'immer';
 
 function reducer(state, action) {
+  console.log('일해라');
   switch (action.type) {
     case 'LOADING':
       return {
@@ -21,7 +22,6 @@ function reducer(state, action) {
         error: action.error,
       };
     case 'NEWCOMMENT':
-      console.log(action);
       return produce(state, draft => {
         draft.data.commentList.unshift({
           content: action.content,
